Integrate LLVM at llvm/llvm-project@8bd65e5
Updates LLVM usage to match llvm/llvm-project@8bd65e5.
* TOSA ops switched to DenseArrayAttr with commit llvm/llvm-project@9e1a344.
  The types of attributes are therfore aligned in the reference
  implementation.
* Uses std::nullopt instead of None.
* Moves from llvm::makeArrayRef to ArrayRef deduction guides, see
  https://reviews.llvm.org/D140896.

Further updates the MLIR-HLO submodule to tensorflow/mlir-hlo@78c6dcf
and bumps TensorFlow to tf-nightly 2.12.0.dev20230112.

+++ reference-implementation/include/emitc/tosa.h
@@ -112,7 +112,7 @@ inline Src reciprocal(Src x) {
 template <typename Dest, size_t Dim, typename Src>
 inline Dest rescale(Src x, typename get_element_type<Src>::type in_zp,
                     typename get_element_type<Dest>::type out_zp,
-                    Tensor1D<int64_t, Dim> mult, Tensor1D<int64_t, Dim> shift,
+                    Tensor1D<int32_t, Dim> mult, Tensor1D<int32_t, Dim> shift,
                     bool scale32, bool double_round, bool per_channel) {
   using ET_Dest = typename get_element_type<Dest>::type;
   using Dest_I32 = typename replace_element_type<int32_t, Dest>::type;
@@ -758,7 +758,7 @@ inline Dest pad(Src operand, Padding padding,
 // TileOp
 // Overload for 1d case
 template <typename Dest, typename Src, IsTensorOfDim<1, Dest> = true>
-Dest tile(Src input, Tensor1D<int32_t, 1> multiples) {
+Dest tile(Src input, Tensor1D<int64_t, 1> multiples) {
   Dest result;
   auto it = result.begin();
   for (int32_t i = 0, M0 = multiples[0]; i < M0; i++) {
@@ -769,7 +769,7 @@ Dest tile(Src input, Tensor1D<int32_t, 1> multiples) {
 
 // Overload for 2d case
 template <typename Dest, typename Src, IsTensorOfDim<2, Src> = true>
-Dest tile(Src input, Tensor1D<int32_t, 2> multiples) {
+Dest tile(Src input, Tensor1D<int64_t, 2> multiples) {
   Dest result;
   auto it = result.begin();
   for (int32_t i = 0, M0 = multiples[0]; i < M0; i++) {
@@ -786,7 +786,7 @@ Dest tile(Src input, Tensor1D<int32_t, 2> multiples) {
 
 // Overload for 3d case
 template <typename Dest, typename Src, IsTensorOfDim<3, Src> = true>
-Dest tile(Src input, Tensor1D<int32_t, 3> multiples) {
+Dest tile(Src input, Tensor1D<int64_t, 3> multiples) {
   Dest result;
   auto it = result.begin();
   for (int32_t m0 = 0, M0 = multiples[0]; m0 < M0; m0++) {
@@ -807,7 +807,7 @@ Dest tile(Src input, Tensor1D<int32_t, 3> multiples) {
 
 // Overload for 4d case
 template <typename Dest, typename Src, IsTensorOfDim<4, Src> = true>
-Dest tile(Src input, Tensor1D<int32_t, 4> multiples) {
+Dest tile(Src input, Tensor1D<int64_t, 4> multiples) {
   Dest result;
   auto it = result.begin();
   for (int32_t m0 = 0, M0 = multiples[0]; m0 < M0; m0++) {
